==> 1_Step(final)_Team3.ipynb <==
## Imports and Defines

! pip install -q gdown

import sqlite3
import os
import pandas as pd
import gdown
import warnings

warnings.filterwarnings('ignore')

db_name = 'database.sqlite'
db_folder = 'src'
if not os.path.exists(db_folder):
    os.mkdir(db_folder)

db_path = os.path.join(db_folder, db_name)

if not os.path.exists(db_path):
    url_db = f'https://drive.google.com/uc?id=1Ez76povTEsoi9ZJj0I8swdczV4F3GQiF'
    gdown.download(url_db, db_path)

con = sqlite3.connect(db_path)


def req(query, connect=con):
    return pd.read_sql_query(query, connect)

## In Process

Посмотрим, какие таблицы есть в Базе Данных

req('''SELECT name FROM sqlite_schema WHERE type='table' AND name NOT LIKE 'sqlite_%';''')

players = req('''
SELECT Player.player_api_id,
       player_name,
       birthday,
       date,
       overall_rating,
       potential,
       preferred_foot,
       attacking_work_rate,
       defensive_work_rate,
       crossing,
       finishing,
       heading_accuracy,
       short_passing,
       volleys,
       dribbling,
       curve,
       free_kick_accuracy,
       long_passing,
       ball_control,
       acceleration,
       sprint_speed,
       agility,
       reactions,
       balance,
       shot_power,
       jumping,
       stamina,
       strength,
       long_shots,
       aggression,
       interceptions,
       positioning,
       vision,
       penalties,
       marking,
       standing_tackle,
       sliding_tackle,
       gk_diving,
       gk_handling,
       gk_kicking,
       gk_positioning,
       gk_reflexes
FROM Player_Attributes
         LEFT JOIN Player ON Player_Attributes.player_api_id = Player.player_api_id;
''')
players.to_csv('./src/players.csv', index=False)
players.head(5)

teams = req('''
SELECT *
FROM Team_Attributes
         LEFT JOIN Team USING (team_api_id);
''')
teams.to_csv('./src/teams.csv', index=False)
teams.info(5)

matches = req('''
SELECT Match.id,
       Country.name AS country_name,
       League.name  AS league_name,
       season,
       stage,
       date,
       card,
       match_api_id,
       home_team_api_id,
       away_team_api_id,
       home_team_goal,
       away_team_goal,
       home_player_1,
       home_player_2,
       home_player_3,
       home_player_4,
       home_player_5,
       home_player_6,
       home_player_7,
       home_player_8,
       home_player_9,
       home_player_10,
       home_player_11,
       away_player_1,
       away_player_2,
       away_player_3,
       away_player_4,
       away_player_5,
       away_player_6,
       away_player_7,
       away_player_8,
       away_player_9,
       away_player_10,
       away_player_11,
       B365H,
       B365D,
       B365A,
       BWH,
       BWD,
       BWA,
       IWH,
       IWD,
       IWA,
       LBH,
       LBD,
       LBA,
       PSH,
       PSD,
       PSA,
       WHH,
       WHD,
       WHA,
       SJH,
       SJD,
       SJA,
       VCH,
       VCD,
       VCA,
       GBH,
       GBD,
       GBA,
       BSH,
       BSD,
       BSA
FROM Match
         JOIN Country ON Match.country_id = Country.id
         JOIN League ON Country.id = League.country_id
''')
matches.to_csv('./src/matches.csv', index=False)
matches.head(5)
