==> 2.5_Step(final)_Team3.ipynb <==
import warnings
warnings.filterwarnings("ignore")
import pandas as pd
import numpy as np
from sklearn.linear_model import LinearRegression
from sklearn.metrics import mean_squared_error, mean_absolute_error, r2_score

df = pd.read_csv('./src/final_dataset.csv')
df.columns

Для прогнозных данных нужны только самые актуальные показатели

df1_home = df[['date', 'country_name', 'league_name', 'team_long_name_home', 'team_short_name_home', 'home_players_average_rating', 'home_players_average_score', 'home_team_score']]
df1_home=df1_home.sort_values('date').drop_duplicates(['country_name', 'league_name',
       'team_long_name_home', 'team_short_name_home'], keep='last')
df1_home.drop('date', axis= 1 , inplace= True )
df1_away = df[['date', 'country_name', 'league_name', 'team_long_name_away', 'team_short_name_away', 'away_players_average_rating', 'away_players_average_score', 'away_team_score']]
df1_away=df1_away.sort_values('date').drop_duplicates(['country_name', 'league_name',
       'team_long_name_away', 'team_short_name_away'], keep='last')
df1_away.drop('date', axis= 1 , inplace= True )

Проверим, каждая ли команда играла и на своем поле, и на поле противника.

print(len(df1_home))
print(len(df1_away))

Определим недостающую команду, заполним недостающие данные для нее и добавим в массив

missed_team=df1_away[ ~ df1_away['team_long_name_away']. isin (df1_home['team_long_name_home'])]
missed_team.rename(columns = {'team_long_name_away':'team_long_name_home', 'team_short_name_away':'team_short_name_home', 'away_players_average_rating':'home_players_average_rating', 'away_players_average_score':'home_players_average_score', 'away_team_score':'home_team_score'}, inplace = True )
df1_home=pd.concat([df1_home, missed_team], ignore_index=True)
len(df1_home)

Соберем все возможные комбинации команд внутри лиг

cross_joined_df = pd.concat(
    [
        sub_df.merge(sub_df, on=['country_name','league_name']).query('team_long_name_home_x != team_long_name_home_y')
        for _, sub_df in df1_home.groupby('league_name')
    ])
cross_joined_df.rename(columns = {'team_long_name_home_x':'team_long_name_home', 'team_short_name_home_x':'team_short_name_home',
                                  'home_players_average_rating_x':'home_players_average_rating', 'home_players_average_score_x':'home_players_average_score', 'home_team_score_x':'home_team_score',
                                 'team_long_name_home_y':'team_long_name_away', 'team_short_name_home_y':'team_short_name_away',
                                 'home_players_average_rating_y':'away_players_average_rating', 'home_players_average_score_y':'away_players_average_score', 'home_team_score_y':'away_team_score'}, inplace = True )
cross_joined_df

Добавим переменные, использованные в модели

cross_joined_df['rating_diff']=cross_joined_df['home_players_average_rating']-cross_joined_df['away_players_average_rating']
cross_joined_df['players_score_diff']=cross_joined_df['home_players_average_score']-cross_joined_df['away_players_average_score']
cross_joined_df['team_diff']=cross_joined_df['home_team_score']-cross_joined_df['away_team_score']
#cross_joined_df['bet_diff']=cross_joined_df['avg_bet_home']-cross_joined_df['avg_bet_away']

Приведем прогноз к формату исторических данных, использованных в основной модели проекта

homedf1=cross_joined_df[['country_name', 'league_name', 'team_long_name_home', 'team_short_name_home',
    'rating_diff', 'players_score_diff', 'team_diff']]
homedf1.rename(columns = {'team_long_name_home':'team_long_name', 'team_short_name_home':'team_short_name'}, inplace = True )
homedf1['homestage']=1
awaydf1=cross_joined_df[['country_name', 'league_name', 'team_long_name_away', 'team_short_name_away',
    'rating_diff', 'players_score_diff', 'team_diff']]
awaydf1.rename(columns = {'team_long_name_away':'team_long_name', 'team_short_name_away':'team_short_name'}, inplace = True )
awaydf1['homestage']=0
awaydf1['rating_diff'] = -1*awaydf1['rating_diff']
awaydf1['team_diff'] = -1*awaydf1['team_diff']
awaydf1['players_score_diff'] = -1*awaydf1['players_score_diff']
cross_joined_df = pd.concat([homedf1, awaydf1], ignore_index=True)
cross_joined_df.columns

Вернемся к историческим данным и спрогнозируем bet_diff, avg_bet_home, avg_bet_away

df['rating_diff']=df['home_players_average_rating']-df['away_players_average_rating']
df['players_score_diff']=df['home_players_average_score']-df['away_players_average_score']
df['team_diff']=df['home_team_score']-df['away_team_score']
df['bet_diff']=df['avg_bet_home']-df['avg_bet_away']

homedf=df[['date', 'season', 'country_name', 'league_name', 'match_api_id', 'team_long_name_home', 'team_short_name_home',
    'match_result', 'rating_diff', 'players_score_diff', 'team_diff', 'bet_diff']]
homedf.rename(columns = {'team_long_name_home':'team_long_name', 'team_short_name_home':'team_short_name'}, inplace = True )
homedf['homestage']=1
homedf['match_result'] = homedf['match_result'].apply(lambda x: 1 if x == 1 else 0)
awaydf=df[['date', 'season', 'country_name', 'league_name', 'match_api_id', 'team_long_name_away', 'team_short_name_away',
    'match_result', 'rating_diff', 'players_score_diff', 'team_diff', 'bet_diff']]
awaydf.rename(columns = {'team_long_name_away':'team_long_name', 'team_short_name_away':'team_short_name'}, inplace = True )

awaydf['homestage']=0
awaydf['match_result'] = awaydf['match_result'].apply(lambda x: 1 if x == -1 else 0)
awaydf['rating_diff'] = -1*awaydf['rating_diff']
awaydf['team_diff'] = -1*awaydf['team_diff']
awaydf['players_score_diff'] = -1*awaydf['players_score_diff']
awaydf['bet_diff'] = -1*awaydf['bet_diff']
df = pd.concat([homedf, awaydf], ignore_index=True)
df.shape

Подберем возможные варианты коэффициентов букмекеров для будущих игр

X = df[['rating_diff', 'players_score_diff', 'team_diff', 'homestage']]
Y = df['bet_diff']

test_split_index = int(X.shape[0]*0.7)
X_train = X[:test_split_index]
X_test = X[test_split_index:]
Y_train = Y[:test_split_index]
Y_test = Y[test_split_index:]

model = LinearRegression()

model.fit(X_train, Y_train)

Y_predicted = model.predict(X_test)
Y_predicted

print('MSE: %.2f' % mean_squared_error(Y_test, Y_predicted))
print('MAE: %.2f' % mean_absolute_error(Y_test, Y_predicted))
print('R2: %.2f' % r2_score(Y_test, Y_predicted))

Вероятность очень высокая, теперь, используя эту модель, мы можем заполнить bet_diff для прогноза 

predicted_bets=model.predict(cross_joined_df[['rating_diff', 'players_score_diff', 'team_diff', 'homestage']])
cross_joined_df['bet_diff']=predicted_bets

cross_joined_df.info()

cross_joined_df[['country_name', 'league_name',
       'team_long_name', 'team_short_name', 'rating_diff', 'players_score_diff', 'team_diff', 'bet_diff', 'homestage']].to_csv(
    './src/new_variables.csv', index=False)
