==> 2.5_Step.ipynb <==
import pandas as pd
import numpy as np
from sklearn.linear_model import LinearRegression
from sklearn.metrics import mean_squared_error, mean_absolute_error, r2_score

df = pd.read_csv('./src/final_dataset.csv')
df.columns

для начала подберем возможные варианты коэффициентов букмекеров для будущих игр

X = df[['home_players_average_rating', 'home_team_prob_case_win',
       'away_players_average_rating', 'away_team_prob_case_win']]
Y = df[['home_avg_bet', 'draw_avg_bet',
       'away_avg_bet']]

test_split_index = int(X.shape[0]*0.7)
X_train = X[:test_split_index]
X_test = X[test_split_index:]
Y_train = Y[:test_split_index]
Y_test = Y[test_split_index:]

model = LinearRegression()

model.fit(X_train, Y_train)

Y_predicted = model.predict(X_test)
bets_df=pd.DataFrame(Y_predicted, columns=['home_avg_bet1', 'draw_avg_bet1',
       'away_avg_bet1'])
bets_df

print('MSE: %.2f' % mean_squared_error(Y_test, Y_predicted))
print('MAE: %.2f' % mean_absolute_error(Y_test, Y_predicted))
print('R2: %.2f' % r2_score(Y_test, Y_predicted))

соберем все возможные комбинации команд внутри лиг

df1 = df[['league_name', 'team_long_name_home',
       'team_short_name_home',]]
df1.columns = ['league_name', 'team_long_name',
       'team_short_name']
df2 = df[['league_name', 'team_long_name_away',
       'team_short_name_away',]]
df2.columns = ['league_name', 'team_long_name',
       'team_short_name']
concat_df = pd.concat([df1, df2], ignore_index=True)
concat_df.drop_duplicates(inplace=True)
cross_joined_df = pd.concat(
    [
        sub_df.merge(sub_df, on='league_name').query('team_long_name_x != team_long_name_y')
        for _, sub_df in concat_df.groupby('league_name')
    ]
)
cross_joined_df.columns = ['league_name', 'team_long_name_home',
       'team_short_name_home','team_long_name_away',
       'team_short_name_away']
cross_joined_df

подтянем известные данные, которые впоследствии будут использованы как независимые переменные

temp_df=df
temp_df['date'] = pd.to_datetime(temp_df['date'])
temp_df=temp_df.sort_values('date').drop_duplicates(['league_name', 'team_long_name_home',
       'team_short_name_home', 'team_short_name_away', 'team_long_name_away',], keep='last')
cross_joined_df=pd.merge(cross_joined_df, temp_df, on=['league_name', 'team_long_name_home',
       'team_short_name_home','team_long_name_away',
       'team_short_name_away'], how='left')

cross_joined_df.info()

заполнились только те строчки, в которых home team и away team уже играли друг с другом на тех же полях. для остальных строк данные нужно подтянуть 2 путями: отдельно для переменных, которые не зависят от того, играет ли команда на своем поле или нет. начнем с players_average_rating и team_score, значения которых не привязаны к параметру home/away.

temp_df

# заполним рейтинги и очки команд для новых комбинаций пар
temp_df1 = temp_df[['date', 'team_long_name_home',
       'team_short_name_home','home_players_average_rating',
       'home_team_score']]
temp_df1.columns = ['date', 'team_long_name',
       'team_short_name','players_average_rating',
       'team_score']
temp_df2 = temp_df[['date', 'team_long_name_away',
       'team_short_name_away','away_players_average_rating',
       'away_team_score']]
temp_df2.columns = ['date', 'team_long_name',
       'team_short_name','players_average_rating',
       'team_score']
concat_temp_df = pd.concat([temp_df1, temp_df2], ignore_index=True)
concat_temp_df=concat_temp_df.sort_values('date').drop_duplicates(['team_long_name', 'team_short_name'], keep='last')
cross_joined_df['home_players_average_rating']=cross_joined_df['home_players_average_rating'].fillna(cross_joined_df.apply(lambda x: concat_temp_df[concat_temp_df['team_long_name'] == x['team_long_name_home'] ]['players_average_rating'].values[0] if pd.isnull(x['home_players_average_rating']) else x['home_players_average_rating'], axis=1))
cross_joined_df['away_players_average_rating']=cross_joined_df['away_players_average_rating'].fillna(cross_joined_df.apply(lambda x: concat_temp_df[concat_temp_df['team_long_name'] == x['team_long_name_home'] ]['players_average_rating'].values[0] if pd.isnull(x['away_players_average_rating']) else x['away_players_average_rating'], axis=1))
cross_joined_df['home_team_score']=cross_joined_df['home_team_score'].fillna(cross_joined_df.apply(lambda x: concat_temp_df[concat_temp_df['team_long_name'] == x['team_long_name_home'] ]['team_score'].values[0] if pd.isnull(x['home_team_score']) else x['home_team_score'], axis=1))
cross_joined_df['away_team_score']=cross_joined_df['away_team_score'].fillna(cross_joined_df.apply(lambda x: concat_temp_df[concat_temp_df['team_long_name'] == x['team_long_name_home'] ]['team_score'].values[0] if pd.isnull(x['away_team_score']) else x['away_team_score'], axis=1))

далее подберем параметры, значения которых зависят от поля (home/away). заполним team_prob_receive_red_cards и team_prob_case_win

temp_df3 = temp_df[['date', 'team_long_name_home',
       'team_short_name_home','home_team_prob_receive_red_cards', 'home_team_prob_case_win']]
temp_df3=temp_df3.sort_values('date').drop_duplicates(['team_long_name_home', 'team_short_name_home'], keep='last')
temp_df3['date'] = temp_df3['date'].dt.year
temp_df4 = temp_df[['date', 'team_long_name_away',
       'team_short_name_away','away_team_prob_receive_red_cards', 'away_team_prob_case_win']]
temp_df4=temp_df4.sort_values('date').drop_duplicates(['team_long_name_away', 'team_short_name_away'], keep='last')
temp_df4['date'] = temp_df4['date'].dt.year
temp_df3

одна из команд никогда не играла на своем поле. нужно найти ее и присвоить значение, исходя из аналогичного параметра для away

missed_team=temp_df4[ ~ temp_df4['team_long_name_away']. isin (temp_df3['team_long_name_home'])]
missed_team

рассчитаем, насколько home_team_prob_receive_red_cards и home_team_prob_case_win больше/меньше away_team_prob_receive_red_cards и away_team_prob_case_win в среднем. 

temp_df3.columns= ['date', 'team_long_name_away',
       'team_short_name_away','home_team_prob_receive_red_cards', 'home_team_prob_case_win']
temp_df3

temp_df3 = pd.merge(temp_df3, temp_df4, on=['date', 'team_long_name_away',
       'team_short_name_away'], how='right')
temp_df3.info()

check_joined_df = temp_df3[(temp_df3.away_team_prob_receive_red_cards != 0 ) & (temp_df3.away_team_prob_case_win != 0 )]
check_joined_df.dropna ()
check_joined_df['prop_red_cards']=check_joined_df['home_team_prob_receive_red_cards']/check_joined_df['away_team_prob_receive_red_cards']
check_joined_df['prop_prob_case_win']=check_joined_df['home_team_prob_case_win']/check_joined_df['away_team_prob_case_win']
print(check_joined_df['prop_red_cards'].mean(skipna=True))
print(check_joined_df['prop_prob_case_win'].mean(skipna=True))

заполним значения home_team_prob_receive_red_cards и home_team_prob_case_win для команды KSV Roeselare, которая никогда не играла на своем поле

temp_df3.columns= ['date', 'team_long_name',
       'team_short_name','home_team_prob_receive_red_cards', 'home_team_prob_case_win','away_team_prob_receive_red_cards', 'away_team_prob_case_win']
temp_df3['home_team_prob_receive_red_cards']=temp_df3['home_team_prob_receive_red_cards'].fillna(temp_df3['away_team_prob_receive_red_cards']*check_joined_df['prop_red_cards'].mean(skipna=True))
temp_df3['home_team_prob_case_win']=temp_df3['home_team_prob_case_win'].fillna(temp_df3['away_team_prob_case_win']*check_joined_df['prop_prob_case_win'].mean(skipna=True))
temp_df3.info()

cross_joined_df['home_team_prob_receive_red_cards']=cross_joined_df['home_team_prob_receive_red_cards'].fillna(cross_joined_df.apply(lambda x: temp_df3[temp_df3['team_long_name'] == x['team_long_name_home'] ]['home_team_prob_receive_red_cards'].values[0] if pd.isnull(x['home_team_prob_receive_red_cards']) else x['home_team_prob_receive_red_cards'], axis=1))
cross_joined_df['away_team_prob_receive_red_cards']=cross_joined_df['away_team_prob_receive_red_cards'].fillna(cross_joined_df.apply(lambda x: temp_df3[temp_df3['team_long_name'] == x['team_long_name_away'] ]['away_team_prob_receive_red_cards'].values[0] if pd.isnull(x['away_team_prob_receive_red_cards']) else x['away_team_prob_receive_red_cards'], axis=1))
cross_joined_df['home_team_prob_case_win']=cross_joined_df['home_team_prob_case_win'].fillna(cross_joined_df.apply(lambda x: temp_df3[temp_df3['team_long_name'] == x['team_long_name_home'] ]['home_team_prob_case_win'].values[0] if pd.isnull(x['home_team_prob_case_win']) else x['home_team_prob_case_win'], axis=1))
cross_joined_df['away_team_prob_case_win']=cross_joined_df['away_team_prob_case_win'].fillna(cross_joined_df.apply(lambda x: temp_df3[temp_df3['team_long_name'] == x['team_long_name_away'] ]['away_team_prob_case_win'].values[0] if pd.isnull(x['away_team_prob_case_win']) else x['away_team_prob_case_win'], axis=1))

predicted_bets=model.predict(cross_joined_df[['home_players_average_rating',
       'home_team_prob_case_win', 'away_players_average_rating',
       'away_team_prob_case_win']])
cross_joined_df[['home_avg_bet', 'draw_avg_bet',
       'away_avg_bet']]=predicted_bets

cross_joined_df.info()

cross_joined_df.drop(['date', 'season', 'match_result', 'home_team_goal', 'home_players_average_score', 'away_team_goal', 'away_players_average_score'], axis= 1 , inplace= True )

cross_joined_df.info()

cross_joined_df.to_csv(
    './src/new_variables.csv', index=False)
