==> 3_Step (1).ipynb <==
import pandas as pd

from sklearn.metrics import accuracy_score, precision_score, recall_score
from sklearn.linear_model import LogisticRegression
from sklearn.model_selection import train_test_split
from sklearn.multioutput import MultiOutputClassifier

df = pd.read_csv('./src/final_dataset.csv')

df.columns

df_encoded = pd.get_dummies(df, columns=['match_result'])

Параметры для обучения

X = df[['home_team_score', 'away_team_score', 'home_players_average_rating', 'away_players_average_rating',
        'avg_bet_home', 'avg_bet_draw',
        'avg_bet_away']]

y = df_encoded[['match_result_Home Win', 'match_result_Away Win', 'match_result_Draw']]

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.25)

log_reg = LogisticRegression(max_iter=500)

multi_target_log_reg = MultiOutputClassifier(log_reg, n_jobs=-1)
multi_target_log_reg.fit(X_train, y_train);

y_predicted = multi_target_log_reg.predict(X_test)

print(f'Accuracy: {accuracy_score(y_test, y_predicted)}')
print(f'Recall: {recall_score(y_test, y_predicted, average="weighted")}')
print(f'Precision: {precision_score(y_test, y_predicted, average="weighted", zero_division=0)}')
