==> 3_Step (2).ipynb <==
import pandas as pd

from sklearn.ensemble import RandomForestClassifier

from sklearn.model_selection import train_test_split
from sklearn.metrics import accuracy_score, precision_score, recall_score

df = pd.read_csv('./src/final_dataset.csv')

df.columns

Параметры для обучения

X = df[['home_team_score', 'away_team_score', 'league_average_rating', 'home_players_average_rating', 'away_players_average_rating',
        'home_players_average_score', 'away_players_average_score', 'avg_bet_home', 'avg_bet_draw',
        'avg_bet_away']]

y = df['match_result']

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.25, random_state=42)

rf_model = RandomForestClassifier(n_estimators=100, random_state=42)
rf_model.fit(X_train, y_train)

y_pred = rf_model.predict(X_test)
y_pred

print("Accuracy:", accuracy_score(y_test, y_pred))
print("Precision:", precision_score(y_test, y_pred, average='macro'))
print("Recall:", recall_score(y_test, y_pred, average='macro'))
