==> 3_Step(final)_Team3.ipynb <==
import pandas as pd
import numpy as np

import warnings
warnings.filterwarnings("ignore")

from sklearn.metrics import accuracy_score, precision_score, recall_score
from sklearn.linear_model import LogisticRegression

df = pd.read_csv('./src/final_dataset.csv')

df.columns

Разделим каждую строчку на 2: одну для home_team, одну для away_team

df['rating_diff']=df['home_players_average_rating']-df['away_players_average_rating']
df['players_score_diff']=df['home_players_average_score']-df['away_players_average_score']
df['team_diff']=df['home_team_score']-df['away_team_score']
df['bet_diff']=df['avg_bet_home']-df['avg_bet_away']

homedf=df[['date', 'season', 'country_name', 'league_name', 'match_api_id', 'team_long_name_home', 'team_short_name_home',
    'match_result', 'rating_diff', 'players_score_diff', 'team_diff', 'bet_diff']]
homedf.rename(columns = {'team_long_name_home':'team_long_name', 'team_short_name_home':'team_short_name'}, inplace = True )
homedf['homestage']=1
homedf['match_result'] = homedf['match_result'].apply(lambda x: 1 if x == 1 else 0)
awaydf=df[['date', 'season', 'country_name', 'league_name', 'match_api_id', 'team_long_name_away', 'team_short_name_away',
    'match_result', 'rating_diff', 'players_score_diff', 'team_diff', 'bet_diff']]
awaydf.rename(columns = {'team_long_name_away':'team_long_name', 'team_short_name_away':'team_short_name'}, inplace = True )

awaydf['homestage']=0
awaydf['match_result'] = awaydf['match_result'].apply(lambda x: 1 if x == -1 else 0)
awaydf['rating_diff'] = -1*awaydf['rating_diff']
awaydf['team_diff'] = -1*awaydf['team_diff']
awaydf['players_score_diff'] = -1*awaydf['players_score_diff']
awaydf['bet_diff'] = -1*awaydf['bet_diff']
df = pd.concat([homedf, awaydf], ignore_index=True)

df.info()

Параметры для обучения

log_reg=LogisticRegression()
X = df[['rating_diff', 'players_score_diff', 'team_diff', 'bet_diff',
       'homestage']]
Y = df['match_result']

test_split_index=int(X.shape[0]*0.75)
X_train=X[:test_split_index]
X_test=X[test_split_index:]
Y_train=Y[:test_split_index]
Y_test=Y[test_split_index:]

log_reg.fit(X_train, Y_train)
Y_predicted=log_reg.predict(X_test)

print('Overall accuracy:', accuracy_score(Y_test, Y_predicted))
print('Overall precision:', precision_score(Y_test, Y_predicted))
print('Overall recall:', recall_score(Y_test, Y_predicted))

forecast_df = pd.read_csv('./src/new_variables.csv')
forecast_df.info()

predicted_match_result=log_reg.predict(forecast_df[['rating_diff', 'players_score_diff', 'team_diff', 'bet_diff',
       'homestage']])
forecast_df['match_result']=predicted_match_result
forecast_df.info()

forecast_df.to_csv(
    './src/final_forecast.csv', index=False)
