==> 3_Step_with_forecast.ipynb <==
import pandas as pd
import numpy as np

from sklearn.ensemble import RandomForestClassifier
from sklearn.model_selection import RandomizedSearchCV
from sklearn.model_selection import GridSearchCV
from sklearn.model_selection import train_test_split
from sklearn.metrics import accuracy_score, precision_score, recall_score

from sklearn.preprocessing import OneHotEncoder
from sklearn.compose import ColumnTransformer
from sklearn.pipeline import Pipeline

df = pd.read_csv('./src/final_dataset.csv')

df.columns

Параметры для обучения

columns_to_test = ['league_name',
                   'home_players_average_rating',
                   'away_players_average_rating',
                   'home_team_score',
                   'away_team_score',
                   'home_team_prob_case_win',
                   'away_team_prob_case_win',
                   'away_team_prob_receive_red_cards',
                   'home_team_prob_receive_red_cards',
                   'home_avg_bet',
                   'draw_avg_bet',
                   'away_avg_bet',
                   ]

Вытащим уникальные лиги

leagues = df['league_name'].unique()

Создадим модель для каждой лиги

param_distributions_for_leagues = {
    'n_estimators': np.arange(100, 501, 50),
    'max_features': ['sqrt', 'log2'],
    'max_depth': np.arange(3, 12)
}

models = {}

for league in leagues:
    league_data = df[df['league_name'] == league]
    X = league_data[columns_to_test].drop(columns=['league_name'])
    y = league_data['match_result']

    X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.25, random_state=42)

    model = RandomForestClassifier(n_estimators=100, random_state=42)
    updated_model = RandomizedSearchCV(estimator=model, param_distributions=param_distributions_for_leagues, n_iter=10, cv=5,
                                       scoring='accuracy', random_state=42, error_score='raise', n_jobs=-1)

    updated_model.fit(X_train, y_train)

    models[league] = updated_model
    best_model = updated_model.best_estimator_
    y_pred = best_model.predict(X_test)

    print(f'Accuracy for {league}:', accuracy_score(y_test, y_pred))

Включение лиги как признака

Предобработка

column_transformer = ColumnTransformer(
    [('league_encoder', OneHotEncoder(), ['league_name'])],
    remainder='passthrough')

pipeline = Pipeline([
    ('preprocessor', column_transformer),
    ('classifier', RandomForestClassifier(n_estimators=100, random_state=42))
])

Параметры для RandomizedSearchCV

param_distributions = {
    'classifier__n_estimators': np.arange(100, 501, 50),
    'classifier__max_features': ['sqrt', 'log2'],
    'classifier__max_depth': np.arange(3, 12)
}

Сам RandomizedSearchCV

random_search = RandomizedSearchCV(estimator=pipeline, param_distributions=param_distributions, n_iter=10, cv=5,
                                   scoring='accuracy', random_state=42, error_score='raise', n_jobs=-1, verbose=1)

X = df[columns_to_test]
y = df['match_result']

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.25, random_state=42)

random_search.fit(X_train, y_train)

best_model = random_search.best_estimator_
y_pred = best_model.predict(X_test)

print('Overall accuracy:', accuracy_score(y_test, y_pred))
print('Overall precision:', precision_score(y_test, y_pred, average='macro'))
print('Overall recall:', recall_score(y_test, y_pred, average='macro'))

forecast_df = pd.read_csv('./src/new_variables.csv')
forecast_df.info()

predicted_match_result=best_model.predict(forecast_df[columns_to_test])
forecast_df['match_result']=predicted_match_result
forecast_df.info()

forecast_df.to_csv(
    './src/final_forecast.csv', index=False)
